Created compilation package and more related files there
