+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/ProcessingManager.java
@@ -1,6 +1,8 @@
 package com.dat3m.dartagnan.program.processing;
 
 import com.dat3m.dartagnan.program.Program;
+import com.dat3m.dartagnan.program.processing.compilation.Compilation;
+
 import org.sosy_lab.common.configuration.Configuration;
 import org.sosy_lab.common.configuration.InvalidConfigurationException;
 import org.sosy_lab.common.configuration.Option;
